# Abolish `yield return` due to performance