--- a/examples/BenchmarkEndToEnd/BenchmarkEndToEnd.csproj
+++ b/examples/BenchmarkEndToEnd/BenchmarkEndToEnd.csproj
@@ -1,0 +1,18 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <Authors>Takuya Takeuchi</Authors>
+    <Description>Example of FaceRecognitionDotNet</Description>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.CommandLineUtils" Version="1.1.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\FaceRecognitionDotNet\FaceRecognitionDotNet.csproj" />
+  </ItemGroup>
+
+</Project>