… examples/BenchmarkEndToEnd/obama-1080p.jpg
… examples/BenchmarkEndToEnd/obama-240p.jpg
… examples/BenchmarkEndToEnd/obama-480p.jpg
… examples/BenchmarkEndToEnd/obama-720p.jpg