`yield retrun` occurs performance issue.
Basically each API does not occur performance issue but return value which are from `yield return` takes much  time to be used.

The following notes are from BenchmarkEndToEnd project.
Code is 

````csharp
        private static void TestEndToEnd(Image image)
        {
            var model = _UseCnn ? Model.Cnn : Model.Hog;
            var faceLocations = _FaceRecognition.FaceLocations(image, model: model);
            var faceLocationCount = faceLocations.Count();

            var faceLandmarks = _FaceRecognition.FaceLandmark(image, faceLocations, model: model);
            var faceLandmarkCount = faceLandmarks.Count();

            var encoding = _FaceRecognition.FaceEncodings(image, faceLocations, model: model);
            var faceEncodingCount = encoding.Count();

            // it could do matching for 1 time
            foreach (var faceEncoding in encoding)
                FaceRecognition.FaceDistance(faceEncoding, faceEncoding);

            foreach (var faceEncoding in encoding)
                faceEncoding.Dispose();
        }
````

## Before abolish

````
Timings at 240p:
 - Face locations, landmark, encoding, distance: 1.2746s (0.78 fps)

Timings at 480p:
 - Face locations, landmark, encoding, distance: 2.4778s (0.40 fps)

Timings at 720p:
 - Face locations, landmark, encoding, distance: 4.5250s (0.22 fps)

Timings at 1080p:
 - Face locations, landmark, encoding, distance: 9.0294s (0.11 fps)
````

## After abolish

````
Timings at 240p:
 - Face locations, landmark, encoding, distance: 0.3358s (2.98 fps)

Timings at 480p:
 - Face locations, landmark, encoding, distance: 0.4672s (2.14 fps)

Timings at 720p:
 - Face locations, landmark, encoding, distance: 0.6926s (1.44 fps)

Timings at 1080p:
 - Face locations, landmark, encoding, distance: 1.1954s (0.84 fps)
````